requirements.txt
Django~=5.2.7
python-dotenv~=1.2.1
psycopg2-binary>=2.9
argon2-cffi>=23.0.0

# Production WSGI Server
gunicorn>=21.0.0

# Redis cache backend (for production)
redis>=5.0.0

# Error monitoring (production)
sentry-sdk[django]>=2.0.0

# REST API & JWT Authentication
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# PDF generation
reportlab>=4.0.0

# Two-Factor Authentication for Admin
django-otp>=1.3.0
qrcode>=7.4.0

# CAPTCHA for registration
django-hcaptcha>=0.2.0
